refactor schemaの整理と関係する処理の修正(#30)
