+++ app/schemas/category.py
@@ -1,0 +1,12 @@
+from pydantic import BaseModel
+
+
+class CategoryBase(BaseModel):
+    text: str
+
+
+class Category(CategoryBase):
+    id: int
+
+    class Config:
+        orm_mode = True
