+++ app/schemas/other.py
@@ -1,0 +1,9 @@
+from pydantic import BaseModel
+
+
+class PredictBase(BaseModel):
+    text: str
+
+
+class CollectCategoriesBase(BaseModel):
+    categories: list[str]
