--- app/util/type.py
@@ -1,9 +1,0 @@
-from pydantic.dataclasses import dataclass
-
-
-@dataclass
-class FeedItem:
-    title: str
-    link: str
-    summary: str | None = None
-    published: str | None = None
